+++ src/pages/SketchesPage/ui/SketchesCard/type.ts
@@ -1,0 +1,9 @@
+import { routesData } from '@config';
+
+export type SketchesCardProps = {
+	id: routesData;
+	screenshot: string;
+	title: string;
+	tags: string[];
+	github: string;
+};